refactored steps
