+++ lib/effects.js
@@ -1,15 +1,6 @@
 const recorder = require('./recorder')
 const { debug } = require('./output')
-const { StepConfig } = require('./step')
 const store = require('./store')
-
-function stepConfig(opts) {
-  return new StepConfig(opts)
-}
-
-module.exports = {
-  stepConfig,
-}
 
 /**
  * @module hopeThat
