+++ lib/recorder.js
@@ -179,7 +179,7 @@ module.exports = {
     }
     if (retry === undefined) retry = true
     if (!running && !force) {
-      return
+      return Promise.resolve()
     }
     tasks.push(taskName)
     debug(chalk.gray(`${currentQueue()} Queued | ${taskName}`))
