+++ lib/utils.js
@@ -531,3 +531,14 @@ module.exports.searchWithFusejs = function (source, searchString, opts) {
 
   return fuse.search(searchString)
 }
+
+module.exports.humanizeString = function (string) {
+  // split strings by words, then make them all lowercase
+  const _result = string
+    .replace(/([a-z](?=[A-Z]))/g, '$1 ')
+    .split(' ')
+    .map(word => word.toLowerCase())
+
+  _result[0] = _result[0] === 'i' ? this.ucfirst(_result[0]) : _result[0]
+  return _result.join(' ').trim()
+}
