+++ lib/within.js
@@ -3,7 +3,7 @@ const store = require('./store')
 const recorder = require('./recorder')
 const container = require('./container')
 const event = require('./event')
-const Step = require('./step')
+const MetaStep = require('./step/meta')
 const { isAsyncFunction } = require('./utils')
 
 /**
@@ -76,7 +76,7 @@ function within(context, fn) {
 
 module.exports = within
 
-class WithinStep extends Step.MetaStep {
+class WithinStep extends MetaStep {
   constructor(locator, fn) {
     super('Within')
     this.args = [locator]
